+++ Engine/Renderer/GLES3/include/GLES3TextureCubeMap.h
@@ -17,7 +17,7 @@ public:
 	uint32_t GetTarget(void) const;
 
 public:
-	bool Create(uint32_t texture);
+	bool Create(uint64_t texture);
 	bool Create(GfxPixelFormat pixelFormat, int width, int height, int levels);
 	void Destroy(void);
