+++ Engine/Renderer/Vulkan/src/VKTexture2D.cpp
@@ -23,7 +23,7 @@ uint32_t CVKTexture2D::GetTarget(void) const
 	return VK_IMAGE_TYPE_2D;
 }
 
-bool CVKTexture2D::Create(uint32_t texture)
+bool CVKTexture2D::Create(uint64_t texture)
 {
 	return false;
 }
